moving logical_reduce from ml_hdl_operations to opt_utils
