+++ metalibm_core/opt/opt_utils.py
@@ -30,9 +30,12 @@
 # last-modified:    Mar  7th, 2018
 ###############################################################################
 
+from metalibm_core.core.ml_formats import ML_Bool
+
 from metalibm_core.core.ml_operations import (
     ML_LeafNode, Comparison, BooleanOperation,
     is_leaf_node,
+    LogicalAnd, LogicalOr,
 )
 from metalibm_core.core.ml_hdl_operations import (
     PlaceHolder
@@ -120,3 +123,21 @@ def depth_node_ordering(start_node, end_nodes):
                 w.append(op)
     return ordered_list
 
+def logical_reduce(op_list, op_ctor=LogicalOr, precision=ML_Bool, **kw):
+    """ Logical/Boolean operand list reduction """
+    local_list = [op for op in op_list]
+    while len(local_list) > 1:
+        op0 = local_list.pop(0)
+        op1 = local_list.pop(0)
+        local_list.append(
+            op_ctor(op0, op1, precision=precision)
+        )
+    # assigning attributes to the resulting node
+    result = local_list[0]
+    result.set_attributes(**kw)
+    return result
+
+## Specialization of logical reduce to OR operation
+logical_or_reduce  = lambda op_list, **kw: logical_reduce(op_list, LogicalOr, ML_Bool, **kw)
+## Specialization of logical reduce to AND operation
+logical_and_reduce = lambda op_list, **kw: logical_reduce(op_list, LogicalAnd, ML_Bool, **kw)
